api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
cachetools==5.3.3
catboost==1.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
filelock==3.13.3
fonttools==4.50.0
fuzzywuzzy==0.18.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
graphviz==0.20.3
greenlet==3.0.3
h11==0.14.0
holidays==0.46
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipykernel==6.29.4
ipython==8.22.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parsel==1.9.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
playwright==1.42.0
plotly==5.20.0
prompt-toolkit==3.0.43
Protego==0.3.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyDispatcher==2.0.7
pyee==11.0.1
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
queuelib==1.6.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
Scrapy==2.11.1
scrapy-playwright==0.0.34
scrapy-user-agents==0.1.1
seaborn==0.13.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tenacity==8.2.3
textdistance==4.6.1
threadpoolctl==3.4.0
tldextract==5.1.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
uvicorn==0.29.0
w3lib==2.1.2
wcwidth==0.2.13
zope.interface==6.2
